fixing fibancci till 98 script

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,10 +8,10 @@
 
 int main(void)
 {
-	double first = 1;
-	double second = 2;
+	long double first = 1;
+	long double second = 2;
 	int numbers = 2;
-	double current;
+	long double current;
 
 	printf("%.0f, %.0f, ", first, second);
 	while (numbers < 98)
